naf/requirements.txt
monai
nibabel
scikit-image
matplotlib
Pillow
tensorboard
gdown
torchVision
tqdm
psutil
pandas
einops
imagecodecs
imgaug
taichi
spams
staintools
timm
opencv-python-headless<4.3
fastremap
#conda install -c conda-forge cupy cudatoolkit=10.2


